feat: 운동 종료 시 행성 경험치 추가 로직 작성

+++ src/main/java/clofi/runningplanet/planet/dto/response/PlanetResponse.java
@@ -6,11 +6,11 @@ public record PlanetResponse(
 	Long planetId,
 	String planetName,
 	String planetImage,
-	Integer demandDistance,
-	Integer distance
+	Double demandDistance,
+	Double distance
 ) {
-	public PlanetResponse(Planet planet, String planetName, String planetImage, Integer demandDistance,
-		Integer distance) {
+	public PlanetResponse(Planet planet, String planetName, String planetImage, Double demandDistance,
+		Double distance) {
 		this(planet.getPlanetImageId(),
 			planetName, planetImage, demandDistance, distance);
 	}
